+++ src/controllers/errors/error.messages.ts
@@ -6,7 +6,9 @@ enum ErrorMessagesForChatRooms {
   incorrectPassword = 'Incorrect password',
   problemCreatingAChatRoom = 'There was a problem trying to create the chat',
   problemGettingAllTheChatsInTheInitialLoad = 'There was a problem trying to get all the chats in the initial load',
-  problemValidatingIfTheUserCanEnter = 'There was a problem trying to verify werther the user can join the chat or not'
+  problemRemovingUserFromChat = 'There was a problem trying to remove the user from the chat room',
+  problemValidatingIfTheUserCanEnter = 'There was a problem trying to verify werther the user can join the chat or not',
+  problemWhileSavingMessage = 'There was a problem trying to save the message.'
 }
 
 export { ErrorMessagesForChatRooms }