+++ src/network/routes.ts
@@ -1,12 +1,11 @@
-import { Application, Router } from 'express'
+import { Application } from 'express'
 import { Home } from '../routes/home'
-import { Users } from '../routes/users'
 
-const routers = [ Users ]
+// const routers = []
 
 const applyRoutes = (app: Application): void => {
   app.use('/', Home)
-  routers.forEach((router: Router): Application => app.use('/api', router))
+  // routers.forEach((router: Router): Application => app.use('/api', router))
 }
 
 export { applyRoutes }